+++ src/Messaging.Sqlite.Server/GenericHandlingServerBuilderExtensions.cs
@@ -1,9 +1,7 @@
-﻿using Assistant.Net.Messaging.Internal;
-using Assistant.Net.Messaging.Options;
+﻿using Assistant.Net.Messaging.Options;
 using Assistant.Net.Options;
 using Assistant.Net.Storage;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
 using System;
 
 namespace Assistant.Net.Messaging;
@@ -21,9 +19,8 @@ public static class GenericHandlingServerBuilderExtensions
     public static GenericHandlingServerBuilder UseSqlite(this GenericHandlingServerBuilder builder, Action<SqliteOptions> configureOptions)
     {
         builder.Services.ConfigureStorage(GenericOptionsNames.DefaultName, b => b
-                .UseSqlite(configureOptions)
-                .UseSqliteSingleProvider())
-            .AddHostedService<ConfigureSqliteHostedService>();
+            .UseSqlite(configureOptions)
+            .UseSqliteSingleProvider());
         return builder;
     }
 
@@ -35,9 +32,8 @@ public static class GenericHandlingServerBuilderExtensions
     public static GenericHandlingServerBuilder UseSqlite(this GenericHandlingServerBuilder builder, string connectionString)
     {
         builder.Services.ConfigureStorage(GenericOptionsNames.DefaultName, b => b
-                .UseSqlite(connectionString)
-                .UseSqliteSingleProvider())
-            .AddHostedService<ConfigureSqliteHostedService>();
+            .UseSqlite(connectionString)
+            .UseSqliteSingleProvider());
         return builder;
     }
 
@@ -49,9 +45,8 @@ public static class GenericHandlingServerBuilderExtensions
     public static GenericHandlingServerBuilder UseSqlite(this GenericHandlingServerBuilder builder, IConfigurationSection configuration)
     {
         builder.Services.ConfigureStorage(GenericOptionsNames.DefaultName, b => b
-                .UseSqlite(configuration)
-                .UseSqliteSingleProvider())
-            .AddHostedService<ConfigureSqliteHostedService>();
+            .UseSqlite(configuration)
+            .UseSqliteSingleProvider());
         return builder;
     }
 }