+++ tests/Storage.Mongo.Tests/SetupMongo.cs
@@ -1,4 +1,5 @@
-﻿using Assistant.Net.Options;
+﻿using Assistant.Net.Abstractions;
+using Assistant.Net.Options;
 using Microsoft.Extensions.DependencyInjection;
 using MongoDB.Bson;
 using MongoDB.Driver;
@@ -11,15 +12,17 @@ namespace Assistant.Net.Storage.Mongo.Tests;
 [SetUpFixture]
 public class SetupMongo
 {
+    private ServiceProvider? provider;
+
     [OneTimeSetUp]
     public async Task OneTimeSetUp()
     {
-        await using var provider = new ServiceCollection()
+        provider = new ServiceCollection()
             .AddMongoClient()
             .ConfigureMongoOptions(ConfigureMongo)
             .BuildServiceProvider();
 
-        var client = provider.CreateScope().ServiceProvider.GetRequiredService<IMongoClient>();
+        var client = provider.GetRequiredService<IMongoClient>();
 
         string pingContent;
         try
@@ -40,7 +43,20 @@ public class SetupMongo
             Assert.Ignore($"The tests require mongodb instance at {ConnectionString}.");
     }
 
-    public static void ConfigureMongo(MongoOptions options) => options.Connection(ConnectionString).Database("test");
+    [OneTimeTearDown]
+    public async Task OneTimeTearDown()
+    {
+        if (provider == null)
+            return;
+
+        var options = provider.GetRequiredService<INamedOptions<MongoOptions>>().Value;
+        var client = provider.GetRequiredService<IMongoClient>();
+        await client.DropDatabaseAsync(options.DatabaseName, CancellationToken);
+
+        await provider.DisposeAsync();
+    }
+
+    public static void ConfigureMongo(MongoOptions options) => options.Connection(ConnectionString).Database("test").EnsureCreated();
 
     private const string ConnectionString = "mongodb://127.0.0.1:27017";