+++ tests/Storage.Sqlite.Tests/Internal/SqliteHistoricalStorageIntegrationTests.cs
@@ -298,8 +298,8 @@ public class SqliteHistoricalStorageIntegrationTests
     [OneTimeTearDown]
     public void OneTimeTearDown() => Provider?.Dispose();
 
-    [TearDown]
-    public async Task TearDown()
+    [SetUp]
+    public async Task Cleanup()
     {
         var dbContext = await Provider!.GetRequiredService<IDbContextFactory<StorageDbContext>>().CreateDbContextAsync(CancellationToken);
         dbContext.HistoricalKeys.RemoveRange(dbContext.HistoricalKeys);