+++ tests/Storage.Sqlite.Tests/Internal/SqlitePartitionedStorageIntegrationTests.cs
@@ -195,8 +195,8 @@ public class SqlitePartitionedStorageIntegrationTests
     [OneTimeTearDown]
     public void OneTimeTearDown() => Provider?.Dispose();
 
-    [TearDown]
-    public async Task TearDown()
+    [SetUp]
+    public async Task Cleanup()
     {
         var dbContext = await Provider!.GetRequiredService<IDbContextFactory<StorageDbContext>>().CreateDbContextAsync(CancellationToken);
         dbContext.HistoricalKeys.RemoveRange(dbContext.HistoricalKeys);