+++ tests/Storage.Sqlite.Tests/Internal/SqliteStorageIntegrationTests.cs
@@ -235,8 +235,8 @@ public class SqliteStorageIntegrationTests
     [OneTimeTearDown]
     public void OneTimeTearDown() => Provider?.Dispose();
 
-    [TearDown]
-    public async Task TearDown()
+    [SetUp]
+    public async Task Cleanup()
     {
         var dbContext = await Provider!.GetRequiredService<IDbContextFactory<StorageDbContext>>().CreateDbContextAsync(CancellationToken);
         dbContext.StorageKeys.RemoveRange(dbContext.StorageKeys);