+++ tests/Storage.Sqlite.Tests/Internal/SqliteStorageProviderIntegrationTests.cs
@@ -202,8 +202,8 @@ public class SqliteStorageProviderIntegrationTests
     [OneTimeTearDown]
     public void OneTimeTearDown() => Provider?.Dispose();
 
-    [TearDown]
-    public async Task TearDown()
+    [SetUp]
+    public async Task Cleanup()
     {
         var dbContext = await Provider!.GetRequiredService<IDbContextFactory<StorageDbContext>>().CreateDbContextAsync(CancellationToken);
         dbContext.StorageKeys.RemoveRange(dbContext.StorageKeys);